refactor(cli): registration with macros
